Split up event controls components into separate files
