+++ client/app/components/conference_room/layout/controls/NoEventControls.jsx
