+++ client/app/components/conference_room/layout/controls/event/Unstarted.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Button } from 'react-bootstrap';
+
+const Unstarted = ({ onConfirm, onCancel }) => (
+  <div className="event-controls">
+    <Button bsStyle="primary" bsSize="large" onClick={onConfirm}>Confirm</Button>
+    <Button bsStyle="danger" bsSize="large" onClick={onCancel} disabled>Cancel</Button>
+  </div>
+);
+
+Unstarted.propTypes = {
+  onConfirm: React.PropTypes.func,
+  onCancel: React.PropTypes.func
+};
+
+export default Unstarted;
